Fix type error causing 'Unknown failure' Problem is that MBEDTLS_ERR_SSL_WANT_READ is negative number, being coverted to unsigned length value and passed into Napi::Buffer::Copy().

+++ src/DtlsSocket.cc
@@ -46,7 +46,7 @@ Napi::Value DtlsSocket::ReceiveDataFromNode(const Napi::CallbackInfo& info) {
 
 	unsigned char buf[RECV_BUF_LENGTH];
 	memset(buf, 0, RECV_BUF_LENGTH);
-	size_t len = receive_data(buf, RECV_BUF_LENGTH);
+	int len = receive_data(buf, RECV_BUF_LENGTH);
 
 	return len > 0 ? Napi::Buffer<unsigned char>::Copy(env, buf, len) : env.Undefined();
 }
